fix: match ID types in `select_examples`
It's possible that a given data file has integers as ID. RSMTool will read them
as integers and not strings. In that case, we need to make sure that any sample
IDs specified in the configuration file are also converted to the same data type
as the featureset IDs before being compared.