feat: add custom intermediate file descriptions
Add a new dictionary to `utils/constants.py` that contains custom descriptions for the
intermediate files generated by `rsmtool` and `rsmsummarize`. We don't add an entry for
`rsmeval` since the files are a strict subset of `rsmtool`'s files.

We handle subgroups by setting a "ZZZ" placeholder which we can replace
with the subgroup name when processing the description.